Lesson6/requirements.txt
aiogram==3.22.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1